Header, links, CSS styles and data table
Updated the Slammers Beach Volleyball from a h1 to a header format. Also
included a resized png file of the volleyball to accompany the header.

Updated the navigation bar to show which link is currently active and
which link is selected when the mouse is hovering over it.

Updated the competition ladder data table to have a nicer format, with
bold headers and even padding for text in data cells.

Commenced modifications on the main styles CSS file to be used with all
web pages. Currently has formats for headers, images, lists, labels,
tables, web forms and main body.

Miscellaneous changes have been made to each web page.